+++ Changelog.md
@@ -9,8 +9,8 @@ Until then, its only a log of what I've changed to prepare update info on Nexus.
 ----
 
 # 2023.08.20 - 1.1.0
-- Fixed: tbbmalloc / low ram - now share same text (partialy) TODO
-
+- Fixed: tbbmalloc / low ram - now share same text (partialy)
+- Fixed: regex handling for certain SKSE occourences/detections
 
 
 # 2023.07.10 - 1.0.1