+++ Changelog.md
@@ -20,6 +20,7 @@ Until then, its only a log of what I've changed to prepare update info on Nexus.
 - Added: (basic) Check for missing masters
 - Added: Possible texture fix - Cathedral Assets Optimizer 
 - Added: 2 Possible ImprovedCamera solutions
+- Added: Very basic possible "unbound" fix
 
 # 2023.07.10 - 1.0.1
 - Fixed: actual RAM values & analysis display formating